# AI_Range/aigatewayservice/openapi/ai-gateway.yaml
openapi: 3.0.3
info:
  title: AI-Range AI Gateway Service
  version: 2.0.0
  description: >
    REST surface for the AI Gateway Service. Two independent surfaces live
    at the same /v1 base path, distinguished by which bearer token pool
    authenticates them (see securitySchemes below):

    - **Range automation** (`/recommendations`, `/assessments`, `/reports*`,
      `/admin/*`): the range control plane, the event-bus adapter (see
      src/eventAdapter.js), and the instructor dashboard call into the AI
      decision engine's four structured modes, authenticated with a named
      service token (`serviceAuth`).
    - **Student chat** (`/models`, `/chat/completions`, `/embeddings`): an
      OpenAI-compatible surface Open WebUI calls on behalf of a student,
      authenticated with a per-enclave tenant token (`chatAuth`) plus
      student identity forwarded in `X-OpenWebUI-User-*` headers. This is
      the only path allowed to reach the tower's vLLM instances, which bind
      loopback-only - see docs/range-inference-tower.html, "The shape of
      the system".
servers:
  - url: http://localhost:8080/v1
paths:
...
  /recommendations:
    post:
      summary: Request a live-adaptation recommendation
      security: [{ serviceAuth: [] }]
      description: >
        Synchronous call into the AI decision engine's live-adaptation mode.
        Returns a structured recommendation (never free text) produced via
        tool-forced output and validated server-side. Low-latency budget —
        callers should expect low single-digit-second responses.
      operationId: createRecommendation
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string }
          description: >
            If supplied, a retried request with the same key returns the
            cached response instead of triggering a second model call.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationRequest"
      responses:
        "200":
          description: Recommendation generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /assessments:
    post:
      summary: Request a qualitative assessment for one completed objective
      description: >
        Synchronous call into the AI decision engine's assessment mode, used
        by the scoring engine's AI qualitative assessor. quality_score is a
        capped 0-1 signal; the caller's rubric decides how much it's worth.
      operationId: createAssessment
      security: [{ serviceAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentRequest"
      responses:
        "200":
          description: Assessment generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports:
    post:
      summary: Start an after-action report generation job (EP4)
      description: >
        Asynchronous — report generation has no real-time latency budget.
        Returns immediately with a job id; poll GET /reports/{jobId} for
        the result.
      operationId: createReport
      security: [{ serviceAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "202":
          description: Report job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportJobAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports/{jobId}:
    get:
      summary: Poll an after-action report job
      operationId: getReport
      security: [{ serviceAuth: [] }]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Job status/result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------
  # Student chat (Open WebUI). See src/chatProxy.js. All three require
  # chatAuth: a per-enclave tenant bearer token, plus student identity
  # forwarded by Open WebUI in X-OpenWebUI-User-* headers (documented on
  # the security scheme below, not modeled as OpenAPI parameters since
  # they're set by Open WebUI itself, never by a hand-written client).
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    serviceAuth:
      type: http
      scheme: bearer
      description: >
        Named service token (range control plane, instructor dashboard) -
        config.services in src/config.js. AI_GATEWAY_API_KEY is the
        "control-plane" token, kept for backward compatibility with the
        single-shared-secret setup this started as.
    chatAuth:
      type: http
      scheme: bearer
      description: >
        Static per-enclave tenant token (config.tenants) identifying which
        Open WebUI instance a request came from. Student identity is
        additionally required via forwarded headers, set by Open WebUI when
        ENABLE_FORWARD_USER_INFO_HEADERS=true - not modeled here as OpenAPI
        parameters since a hand-written client never sets them directly:
          X-OpenWebUI-User-Id (required), X-OpenWebUI-User-Email,
          X-OpenWebUI-User-Name, X-OpenWebUI-User-Role.
  responses:
    BadRequest:
      description: Request failed schema validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing/malformed Authorization header
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_request }
            message: { type: string }
          required: [code, message]
      required: [error]
    StateSnapshot:
      type: object
      description: Same shape as the State Manager's exercise state snapshot.
      required: [exercise_id, objectives]
      properties:
        exercise_id: { type: string }
        elapsed_minutes: { type: number }
        objectives:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              status: { type: string, enum: [not_started, in_progress, complete] }
              blocked_minutes: { type: number }
        student_actions_last_5min: { type: number }
    RecommendationRequest:
      type: object
      required: [exercise_id, snapshot]
      properties:
        exercise_id: { type: string }
        student_id: { type: string }
        snapshot: { $ref: "#/components/schemas/StateSnapshot" }
        scenario_constraints:
          type: object
          description: Authored objectives/injects/branches this recommendation may reference.
        recent_recommendations:
          type: array
          items: { type: object }
          description: Last few recommendations, to avoid repeating a rejected suggestion.
    Recommendation:
      type: object
      required: [recommendation_type, confidence, rationale]
      properties:
        recommendation_type:
          type: string
          enum: [inject_hint, difficulty_adjust, branch_change, extend_time, no_action]
        target_id:
          type: string
          description: Required unless recommendation_type is no_action.
        confidence: { type: number, minimum: 0, maximum: 1 }
        rationale: { type: string }
    RecommendationResponse:
      type: object
      properties:
        exercise_id: { type: string }
        student_id: { type: string }
        recommendation: { $ref: "#/components/schemas/Recommendation" }
        prompt_version: { type: string }
        schema_version: { type: string }
        attempts: { type: integer }
        generated_at: { type: string, format: date-time }
    AssessmentRequest:
      type: object
      required: [exercise_id, objective_id, action_log]
      properties:
        exercise_id: { type: string }
        student_id: { type: string }
        objective_id: { type: string }
        dimension: { type: string, enum: [methodology, efficiency, documentation] }
        action_log:
          type: array
          items: { type: object }
        submission_text: { type: string }
    Assessment:
      type: object
      required: [objective_id, quality_score, dimension, rationale]
      properties:
        objective_id: { type: string }
        quality_score: { type: number, minimum: 0, maximum: 1 }
        dimension: { type: string, enum: [methodology, efficiency, documentation] }
        evidence: { type: string }
        rationale: { type: string }
    AssessmentResponse:
      type: object
      properties:
        exercise_id: { type: string }
        student_id: { type: string }
        assessment: { $ref: "#/components/schemas/Assessment" }
        prompt_version: { type: string }
        schema_version: { type: string }
        attempts: { type: integer }
        generated_at: { type: string, format: date-time }
    ReportRequest:
      type: object
      required: [exercise_id]
      properties:
        exercise_id: { type: string }
        mode: { type: string, enum: [post_exercise_report] }
        state_snapshots:
          type: array
          items: { $ref: "#/components/schemas/StateSnapshot" }
        recommendations:
          type: array
          items: { $ref: "#/components/schemas/Recommendation" }
        assessments:
          type: array
          items: { $ref: "#/components/schemas/Assessment" }
        final_scores: { type: object }
    ReportJobAccepted:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        status: { type: string, enum: [pending] }
        poll_url: { type: string }
    ReportJob:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, complete, failed] }
        result:
          type: object
          nullable: true
          properties:
            narrative: { type: string }
            promptVersion: { type: string }
            schemaVersion: { type: string }
        error: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
...